add sudo check for scripts. add https properties for production. change api call in frontend to https. add cross origin for frontend

+++ scripts/build_and_deploy_prod.sh
@@ -1,11 +1,20 @@
 #!/bin/bash
 
+if [ "$EUID" -ne 0 ]
+  then echo "Please run as root"
+  exit
+fi
+
 echo "###Docker compose down for webserver_homepage###"
 docker compose --file /home/crypta/dockerContainers/webserver_homepage/docker-compose-fullstack.yml down
 echo "###Done###"
 
 echo "###Building Backend###"
 mvn -f /home/crypta/git/homepage/backend/pom.xml -Dmaven.test.skip clean install
+echo "###Done###"
+
+echo "###Building Frontend###"
+npm --prefix /home/crypta/git/homepage/frontend run build
 echo "###Done###"
 
 echo "###Removing old files###"
